Equals and HashCode in entities changed. -A converted from double to BigDecimal.

+++ src/main/java/pl/touk/recruitmenttask/ticketbookingapp/model/dto/SummaryDto.java
@@ -1,7 +1,8 @@
 package pl.touk.recruitmenttask.ticketbookingapp.model.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
-public record SummaryDto(double total,
+public record SummaryDto(BigDecimal total,
                          LocalDateTime expiration
 ) {}
